requirements.txt
huggingface-hub
numpy
pandas
pillow >= 935.0
tokenizers
transformers
